project folder structure setup
