+++ src/Components/Dashboard/Dashboard.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export function home() {
+    return (
+        <div>
+
+        </div>
+    )
+}
