--- src/HomeScreen/home.js
@@ -1,15 +1,0 @@
-import React, {Component} from 'react'
-import Nav from '../Components/Navgation/Nav'
-class home extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {  };
-    }
-    render() {
-        return (
-            <Nav/>
-        );
-    }
-}
-
-export default home;
